Modularise tests into seporate files. Related #5 and #10
